+++ AlphaFS.UnitTest/DirectoryInfo Class/DirectoryInfo_Junction.cs
@@ -69,7 +69,7 @@ namespace AlphaFS.UnitTest
 
             var lvi = Alphaleonis.Win32.Filesystem.Directory.GetLinkTargetInfo(dirInfo.FullName);
             UnitTestConstants.Dump(lvi, -14);
-            UnitTestConstants.Dump(dirInfo.EntryInfo, -17);
+            UnitTestConstants.Dump(dirInfo.EntryInfo, -19);
 
             Assert.AreEqual(System.IO.Directory.Exists(dirInfo.FullName), Alphaleonis.Win32.Filesystem.Directory.Exists(dirInfo.FullName));
             Assert.AreEqual(junction, dirInfo.FullName);